examples/todo/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
attrs==22.1.0
autoflake==1.7.7
autopep8==2.0.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.4
exceptiongroup==1.0.4
Faker==12.0.1
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mixer==7.2.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
promise==2.3
psycopg2==2.8.6
pycodestyle==2.10.0
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.10.0
